requirements.txt
reflex==0.4.5
Werkzeug==3.0.2
blinker==1.7.0
flask==3.0.3
itsdangerous==2.2.0
Pillow==10.3.0
google-generativeai==0.5.2
python-dotenv==1.0.1
spotipy==2.23.0